+++ ys_codebase/source/server/server/__init__.py
@@ -3,6 +3,7 @@ Server Package - Persistent Daemon & Process Supervisor.
 """
 
 from server.master import MasterSupervisor, ServerDaemonState
+from server.logger import ServerLogger
 from server.worker import WarmWorker
 from server.service import BaseServiceWorker, ServiceManager
 from server.streamer import DebouncedIOStreamer
@@ -12,6 +13,7 @@ __version__ = "1.0.0"
 __all__ = [
     "MasterSupervisor",
     "ServerDaemonState",
+    "ServerLogger",
     "WarmWorker",
     "BaseServiceWorker",
     "ServiceManager",